Refactor move updateMarkets
