+++ subgraphs/euler/src/mappings/handlers.ts
@@ -54,6 +54,7 @@ import { updateFinancials, updateMarketDailyMetrics, updateMarketHourlyMetrics, 
 import { _MarketUtility } from "../../generated/schema";
 import { BigInt } from "@graphprotocol/graph-ts";
 import { getAssetTotalSupply } from "../common/tokens";
+import { updateMarkets } from "./helpers";
 
 export function handleAssetStatus(event: AssetStatus): void {
   const tokenAddress = event.params.underlying.toHexString();
@@ -276,159 +277,6 @@ export function handleLiquidation(event: Liquidation): void {
     }
   }
   liquidation.save();
-}
-
-export function handleMarketActivated(event: MarketActivated): void {
-  const market = getOrCreateMarket(event.params.underlying.toHexString());
-  const marketUtility = getOrCreateMarketUtility(event.params.underlying.toHexString());
-
-  market.createdTimestamp = event.block.timestamp;
-  market.createdBlockNumber = event.block.number;
-
-  // Market are initialized in isolated tier, which means currency can't be used as collateral.
-  // https://docs.euler.finance/risk-framework/tiers
-  market.canUseAsCollateral = false;
-  market.canBorrowFrom = true;
-
-  const token = getOrCreateToken(event.params.underlying);
-  const dToken = getOrCreateToken(event.params.dToken);
-  const eToken = getOrCreateToken(event.params.eToken);
-
-  market.outputToken = eToken.id;
-  market.inputToken = token.id;
-  market.save();
-
-  marketUtility.token = token.id;
-  marketUtility.eToken = eToken.id;
-  marketUtility.dToken = dToken.id;
-  marketUtility.save();
-
-  const protocolUtility = getOrCreateProtocolUtility();
-  protocolUtility.markets = protocolUtility.markets.concat([market.id]);
-  protocolUtility.save();
-}
-
-function updateMarkets(eulerViewQueryResponse: EulerGeneralView__doQueryResultRStruct, block: ethereum.Block): void {
-  // let ethUsdcExchangeRate: BigDecimal;
-  let ethUsdcExchangeRate = BIGDECIMAL_ONE;
-  const eulerViewMarkets = eulerViewQueryResponse.markets;
-
-  // Try to get ETH/USDC exchange rate directly from Euler query...
-  const usdcMarketIndex = eulerViewMarkets.findIndex(m => m.symbol === USDC_SYMBOL);
-  if (usdcMarketIndex !== -1) {
-    ethUsdcExchangeRate = eulerViewMarkets[usdcMarketIndex].currPrice.toBigDecimal();
-  } else {
-    // ...otherwise fallback to Uniswap,
-    // Convert USD/ETH exchange rate to ETH/USD
-    ethUsdcExchangeRate = BIGDECIMAL_ONE.div(getEthPriceUsd());
-  }
-
-  const protocol = getOrCreateLendingProtocol();
-  protocol.totalValueLockedUSD = BIGDECIMAL_ZERO;
-  protocol.totalBorrowBalanceUSD = BIGDECIMAL_ZERO;
-  protocol.totalDepositBalanceUSD = BIGDECIMAL_ZERO;
-
-  // Using an indexed for loop because AssemblyScript does not support closures.
-  // AS100: Not implemented: Closures
-  for (let i = 0; i < eulerViewMarkets.length; i += 1) {
-    const eulerViewMarket = eulerViewMarkets[i];
-    const market = getOrCreateMarket(eulerViewMarket.underlying.toHexString());
-    const lendingRate = getOrCreateInterestRate(InterestRateSide.LENDER, InterestRateType.VARIABLE, market.id);
-    const borrowRate = getOrCreateInterestRate(InterestRateSide.BORROWER, InterestRateType.VARIABLE, market.id);
-    lendingRate.rate = eulerViewMarket.supplyAPY.toBigDecimal().div(INTEREST_RATE_PRECISION);
-    borrowRate.rate = eulerViewMarket.borrowAPY.toBigDecimal().div(INTEREST_RATE_PRECISION);
-    lendingRate.save();
-    borrowRate.save();
-    market.rates = market.rates.concat([lendingRate.id, borrowRate.id]);
-
-    const currPrice = eulerViewMarket.currPrice.toBigDecimal().div(DECIMAL_PRECISION);
-    const currPriceUsd = currPrice.times(ethUsdcExchangeRate);
-
-    const token = getOrCreateToken(eulerViewMarket.underlying);
-    token.lastPriceUSD = currPriceUsd;
-    token.lastPriceBlockNumber = block.number;
-    token.save();
-
-    const tokenPrecision = new BigDecimal(BigInt.fromI32(10).pow(<u8>token.decimals));
-    market.totalDepositBalanceUSD = eulerViewMarket.totalBalances
-      .toBigDecimal()
-      .div(tokenPrecision)
-      .times(currPriceUsd);
-    market.totalBorrowBalanceUSD = eulerViewMarket.totalBorrows
-      .toBigDecimal()
-      .div(tokenPrecision)
-      .times(currPriceUsd);
-    market.totalValueLockedUSD = market.totalDepositBalanceUSD.minus(market.totalBorrowBalanceUSD);
-    market.name = token.name;
-    market.inputTokenBalance = eulerViewMarket.totalBalances;
-    market.inputTokenPriceUSD = currPriceUsd;
-    
-    const marketUtility = getOrCreateMarketUtility(market.id);
-   
-    const previousReserveBalance = marketUtility.reserveBalance;
-
-    /**
-     * The following fields are always equal to 0:
-     * - eulerMarketView.eTokenBalance
-     * - eulerMarketView.eTokenBalanceUnderlying
-     * - eulerMarketView.dTokenBalance
-     *
-     * In order to calculate eToken and dToken price, we need to pull supply off ERC-20 contracts and calculate
-     * token price by using totalDepositBalanceUSD and totalBorrowBalanceUSD.
-     */
-    const eTokenAddress = Address.fromString(marketUtility.eToken);
-    const eToken = getOrCreateToken(eTokenAddress);
-    const eTokenTotalSupply = getAssetTotalSupply(eTokenAddress);
-    const eTokenPrecision = new BigDecimal(BigInt.fromI32(10).pow(<u8>eToken.decimals));
-    if (eTokenTotalSupply.gt(BIGINT_ZERO)) {
-      const eTokenPriceUSD = market.totalDepositBalanceUSD.div(eTokenTotalSupply.toBigDecimal().div(eTokenPrecision));
-      eToken.lastPriceUSD = eTokenPriceUSD;
-      eToken.lastPriceBlockNumber = block.number;
-      eToken.save();
-
-      market.outputTokenPriceUSD = eTokenPriceUSD;
-      market.outputTokenSupply = eTokenTotalSupply;
-      market.exchangeRate = eTokenTotalSupply.toBigDecimal().div(eTokenPrecision).div(eulerViewMarket.totalBalances.toBigDecimal().div(tokenPrecision));
-    } 
-
-    const dTokenAddress = Address.fromString(marketUtility.dToken);
-    const dToken = getOrCreateToken(dTokenAddress);
-    const dTokenTotalSupply = getAssetTotalSupply(dTokenAddress);
-    if (dTokenTotalSupply.gt(BIGINT_ZERO)) {
-      const dTokenPrecision = new BigDecimal(BigInt.fromI32(10).pow(<u8>dToken.decimals));
-      const dTokenPriceUSD = market.totalBorrowBalanceUSD.div(dTokenTotalSupply.toBigDecimal().div(dTokenPrecision));
-      dToken.lastPriceUSD = dTokenPriceUSD;
-      dToken.lastPriceBlockNumber = block.number;
-      dToken.save();
-    }
-
-    market.save();
-
-    const reserveBalanceDiff = eulerViewMarket.reserveBalance.minus(previousReserveBalance);
-    // Ignore case where there was a balance conversion (negative diff) happening at current block.
-    if (reserveBalanceDiff.gt(BIGINT_ZERO)) {
-      const marketRevenueDiffUsd = reserveBalanceDiff
-        .toBigDecimal()
-        .div(eTokenPrecision)
-        .times(eToken.lastPriceUSD!);
-      protocol.cumulativeProtocolSideRevenueUSD = protocol.cumulativeProtocolSideRevenueUSD.plus(marketRevenueDiffUsd);
-      protocol.cumulativeTotalRevenueUSD = protocol.cumulativeTotalRevenueUSD.plus(marketRevenueDiffUsd);
-    }
-    
-    protocol.totalValueLockedUSD = protocol.totalValueLockedUSD.plus(market.totalValueLockedUSD);
-    protocol.totalBorrowBalanceUSD = protocol.totalBorrowBalanceUSD.plus(market.totalBorrowBalanceUSD);
-    protocol.totalDepositBalanceUSD = protocol.totalDepositBalanceUSD.plus(market.totalDepositBalanceUSD);
-
-     marketUtility.market = market.id;
-     marketUtility.twap = eulerViewMarket.twap;
-     marketUtility.twapPeriod = eulerViewMarket.twapPeriod;
-     marketUtility.reserveBalance = eulerViewMarket.reserveBalance;
-     marketUtility.save();
-
-    updateMarketDailyMetrics(block, market.id);
-    updateMarketHourlyMetrics(block, market.id);
-  }
-  protocol.save();
 }
 
 export function handleBlockUpdates(block: ethereum.Block): void {
